made VersionedItem abstract, moved registry methods to Registry
added support for selective version output to versions command
