+++ src/main/java/com/kantenkugel/discordbot/versioncheck/updatehandle/UpdateHandler.java
@@ -1,6 +1,6 @@
 package com.kantenkugel.discordbot.versioncheck.updatehandle;
 
-import com.kantenkugel.discordbot.versioncheck.VersionedItem;
+import com.kantenkugel.discordbot.versioncheck.items.VersionedItem;
 
 @FunctionalInterface
 public interface UpdateHandler
